+++ cmd/console_cmd.go
@@ -165,7 +165,7 @@ func openConsoleAccessKey(ctx *RunContext, creds *storage.RoleCredentials, durat
 	}
 	url := login.GetUrl()
 
-	return utils.HandleUrl(ctx.Cli.UrlAction, ctx.Cli.Browser, url,
+	return utils.HandleUrl(ctx.Settings.UrlAction, ctx.Settings.Browser, url,
 		"Please open the following URL in your browser:\n\n", "\n\n")
 }